requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
azure-search-documents>=11.4.0
openai>=1.0.0
requests>=2.31.0

